GH-518 Replace Vavr with a new set of functional panda-utilities (Resolve #525)

+++ panda-framework/src/main/java/org/panda_lang/framework/design/architecture/module/TypesMap.java
@@ -39,7 +39,7 @@ public interface TypesMap extends Map<String, Type>, ModuleResource {
      *
      * @return the amount of used types
      */
-    int countUsedTypes();
+    long countUsedTypes();
 
     /**
      * Get collection of entries that contains types
